Fix sonarqube issues
