+++ vaadin-services-demo/src/test/java/com/vaadin/services/VaadinServicesDemoApplicationTests.java
@@ -15,13 +15,17 @@
  */
 package com.vaadin.services;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
 
-@SpringBootApplication
-public class VaadinServiceDemoApplication {
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class VaadinServicesDemoApplicationTests {
 
-  public static void main(String[] args) {
-    SpringApplication.run(VaadinServiceDemoApplication.class, args);
+  @Test
+  public void contextLoads() {
   }
+
 }
