+++ vaadin-services/pom.xml
@@ -53,23 +53,25 @@
                         <goals>
                             <goal>check</goal>
                         </goals>
-                        <configuration><rules>
-                            <rule implementation="org.jacoco.maven.RuleConfiguration">
-                                <element>BUNDLE</element>
-                                <limits>
-                                    <limit implementation="org.jacoco.report.check.Limit">
-                                        <counter>INSTRUCTION</counter>
-                                        <value>COVEREDRATIO</value>
-                                        <minimum>1.0</minimum>
-                                    </limit>
-                                    <limit implementation="org.jacoco.report.check.Limit">
-                                        <counter>CLASS</counter>
-                                        <value>MISSEDCOUNT</value>
-                                        <maximum>0</maximum>
-                                    </limit>
-                                </limits>
-                            </rule>
-                        </rules></configuration>
+                        <configuration>
+                            <rules>
+                                <rule implementation="org.jacoco.maven.RuleConfiguration">
+                                    <element>BUNDLE</element>
+                                    <limits>
+                                        <limit implementation="org.jacoco.report.check.Limit">
+                                            <counter>INSTRUCTION</counter>
+                                            <value>COVEREDRATIO</value>
+                                            <minimum>1.0</minimum>
+                                        </limit>
+                                        <limit implementation="org.jacoco.report.check.Limit">
+                                            <counter>CLASS</counter>
+                                            <value>MISSEDCOUNT</value>
+                                            <maximum>0</maximum>
+                                        </limit>
+                                    </limits>
+                                </rule>
+                            </rules>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
